refactor(services): Barrel file renaming
