+++ projects/igniteui-angular/src/lib/select/select.common.ts
@@ -1,5 +1,5 @@
 import { IgxDropDownBaseDirective, IgxDropDownItemBaseDirective } from '../drop-down/public_api';
-import { OverlaySettings } from '../services';
+import { OverlaySettings } from '../services/public_api';
 import { IgxInputDirective } from '../input-group/public_api';
 
 /** @hidden @internal */
