+++ projects/igniteui-angular/src/lib/select/select.component.ts
@@ -15,7 +15,7 @@ import { IgxDropDownComponent } from './../drop-down/drop-down.component';
 import { IgxSelectItemComponent } from './select-item.component';
 import { SelectPositioningStrategy } from './select-positioning-strategy';
 
-import { OverlaySettings, AbsoluteScrollStrategy } from '../services/index';
+import { OverlaySettings, AbsoluteScrollStrategy } from '../services/public_api';
 import { IGX_DROPDOWN_BASE, ISelectionEventArgs, Navigate } from '../drop-down/drop-down.common';
 import { CancelableEventArgs } from '../core/utils';
 import { IgxLabelDirective } from '../directives/label/label.directive';
